requirements.txt
Flask
Flask-Assets
gunicorn
requests
json5
beautifulsoup4
fake-useragent
python-dotenv